My initial solution.
This was generated in a fairly TDD fashion along with the construction of the tests in `bats_tests.sh`.

There are some weird 'g' prefixes on several system calls which are there to force my Mac to use the
GNU version of some of these functions where the GNU behavior is different from the BSD behavior that
I get on a Mac. I'll have to remove those before "publishing" the final version of this. I should also probably
have a note in the README that points out that this won't work "as is" on a Mac, but that it's easy enough
to change so that it does.